+++ contracts/index.ts
@@ -1,2 +1,3 @@
 export * from "./gameCatalog";
+export * from "./roomMatch";
 export * from "./userProfile";